PT-2171: Move countries list getting into a service
