+++ src/VirtoCommerce.Platform.Core/Common/ICountriesService.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace VirtoCommerce.Platform.Core.Common
+{
+    public interface ICountriesService
+    {
+        Task<IList<Country>> GetCountriesAsync();
+    }
+}
