+++ src/VirtoCommerce.Platform.Data/Extensions/ServiceCollectionExtensions.cs
@@ -6,6 +6,7 @@ using Microsoft.Extensions.DependencyInjection;
 using VirtoCommerce.Platform.Caching;
 using VirtoCommerce.Platform.Core.Bus;
 using VirtoCommerce.Platform.Core.ChangeLog;
+using VirtoCommerce.Platform.Core.Common;
 using VirtoCommerce.Platform.Core.Events;
 using VirtoCommerce.Platform.Core.ExportImport;
 using VirtoCommerce.Platform.Core.Localizations;
@@ -13,6 +14,7 @@ using VirtoCommerce.Platform.Core.Notifications;
 using VirtoCommerce.Platform.Core.TransactionFileManager;
 using VirtoCommerce.Platform.Core.ZipFile;
 using VirtoCommerce.Platform.Data.ChangeLog;
+using VirtoCommerce.Platform.Data.Common;
 using VirtoCommerce.Platform.Data.DynamicProperties;
 using VirtoCommerce.Platform.Data.ExportImport;
 using VirtoCommerce.Platform.Data.Localizations;
@@ -56,6 +58,7 @@ namespace VirtoCommerce.Platform.Data.Extensions
             services.AddSingleton<ITranslationDataProvider, ModulesTranslationDataProvider>();
             services.AddSingleton<ITranslationService, TranslationService>();
 
+            services.AddSingleton<ICountriesService, FileSystemCountriesService>();
             services.AddSingleton<IFileSystem, FileSystem>();
             services.AddTransient<IZipFileWrapper, ZipFileWrapper>();
